sanity is only updated on a successful purchase

+++ cellmate/pages/success.js
@@ -1,20 +1,49 @@
-import React, { useState, useEffect } from 'react';
-import Link from 'next/link';
-import { BsBagCheckFill } from 'react-icons/bs';
+import React, { useState, useEffect } from "react";
+import Link from "next/link";
+import { BsBagCheckFill } from "react-icons/bs";
 
-import { useStateContext } from '../context/StateContext';
-import { runFireworks } from '../lib/util';
+import { useStateContext } from "../context/StateContext";
+import { runFireworks } from "../lib/util";
+import { useRouter } from "next/router";
 
 const Success = () => {
-  const { setCartItems, setTotalPrice, setTotalQuantities } = useStateContext();
-  
-  useEffect(() => {
-    localStorage.clear();
-    setCartItems([]);
-    setTotalPrice(0);
-    setTotalQuantities(0);
-    runFireworks();
-  }, []);
+  const router = useRouter();
+  const { setCartItems, setTotalPrice, setTotalQuantities, cartItems } =
+    useStateContext();
+
+  useEffect(async () => {
+    if (!router.isReady) return;
+    if (!router.query?.session_id) return;
+
+    const response = await fetch(`/api/stripe?key=${router.query.session_id}`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
+    if (response.statusCode === 500) return;
+    const data = await response.json();
+
+    if (data.session.payment_status === "paid") {
+      if (cartItems.length) {
+        // update the stock amount in sanity
+        const sanityCheck = await fetch("/api/sanityUpdate", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify(cartItems),
+        });
+      }
+
+      setCartItems([]);
+      setTotalPrice(0);
+      setTotalQuantities(0);
+
+      localStorage.clear();
+      runFireworks();
+    }
+  }, [router.isReady]);
 
   return (
     <div className="success-wrapper">
@@ -27,7 +56,7 @@ const Success = () => {
         <p className="description">
           If you have any questions, please email
           <a className="email" href="mailto:order@example.com">
-           OOgunremi@yahoo.com
+            OOgunremi@yahoo.com
           </a>
         </p>
         <Link href="/">
@@ -37,7 +66,7 @@ const Success = () => {
         </Link>
       </div>
     </div>
-  )
-}
+  );
+};
 
-export default Success
+export default Success;
